Rename licenseKey helpers to more meaningful names
