+++ src/helpers/licenseKeyHelper.ts
