+++ test/helpers/licenseKeyValidator/agpl-v3.spec.ts
