+++ test/helpers/licenseKeyValidator/invalid-agpl-version.spec.ts
