+++ test/helpers/licenseKeyValidator/invalid-key.spec.ts
