+++ test/helpers/licenseKeyValidator/missing-key.spec.ts
