+++ test/helpers/licenseKeyValidator/non-commercial-key.spec.ts
